requirements.txt
streamlit==1.44.1
plotly>=5.0.0
pandas>=2.2.0
numpy>=1.26.0
google-cloud-storage>=2.0.0
altair>=5.0.0,<5.4.0
openpyxl
gspread
matplotlib
requests
beautifulsoup4
lxml
